fix(config): move LLM skip helper to break CodeQL import cycle
has_credentials_for_active_llm_provider lived in llm_credentials with a lazy import of LLMSettings, which CodeQL flags as a config<->llm_credentials cycle. Co-locate it with LLMSettings in app.config instead.
